@@ -11,22 +11,25 @@ authors = [
 maintainers = [
   {name = "Per A. Brodtkorb, FFI"}
 ]
+
+requires-python = ">=3.10"
 dependencies = [
-    "numpy>=1.11.0",
-    "scipy>=1.3.0",
+    "numpy>=1.21.2",
+    "scipy>=1.7.3", #; python_version < \"3.14\"",
+    #"scipy>=1.16.1; python_version >= \"3.14\"",
     "karney>=1.0",
 ]
 
-requires-python = ">=3.9"
 readme = "README.rst"
 license = {text = "Simplified BSD"}
 license_file = "LICENSE.txt"
 classifiers = [
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+    "Programming Language :: Python :: 3.14",
     "Environment :: Console",
     "Intended Audience :: Education",
     "Intended Audience :: Science/Research",
