@@ -1,9 +1,11 @@
+from typing import Any, Optional
+
+from ._common import use_docstring
 from ._info import __doc__ as __doc__  # @UnresolvedImport
-from .util import *  # noqa
-from .rotation import *  # noqa
 from .core import *  # noqa
 from .objects import *  # noqa
-from ._common import use_docstring
+from .rotation import *  # noqa
+from .util import *  # noqa
 
 __version__ = "1.0.2"
 
@@ -11,12 +13,12 @@ _PACKAGE_NAME = __name__
 
 
 @use_docstring(
-    """
-import {0} as {1}
-{1}.test('-q', '--doctest-modules', '--cov={0}', '--disable-warnings')
-""".format(_PACKAGE_NAME, _PACKAGE_NAME[:2])
+    f"""
+import {_PACKAGE_NAME} as {_PACKAGE_NAME[:2]}
+{_PACKAGE_NAME[:2]}.test('-q', '--doctest-modules', '--cov={_PACKAGE_NAME}', '--disable-warnings')
+"""
 )
-def test(*options, plugins=None):
+def test(*options: str, plugins: Optional[Any] = None) -> int:
     """
     Run tests for module using pytest.
 
@@ -37,7 +39,7 @@ def test(*options, plugins=None):
 
     Returns
     -------
-    exit_code: scalar
+    exit_code: int
         Exit code is 0 if all tests passed without failure.
 
     Examples
