@@ -1,10 +1,10 @@
 from nvector import (
-    _intro,
-    _documentation,
-    _installation,
-    _examples_object_oriented,
     _acknowledgements,
+    _documentation,
+    _examples_object_oriented,
     _images,
+    _installation,
+    _intro,
 )
 
 __doc__ = (  # @ReservedAssignment
@@ -24,7 +24,7 @@ __doc__ = (  # @ReservedAssignment
 
 
 if __name__ == "__main__":
-    from nvector._common import write_readme, test_docstrings
+    from nvector._common import test_docstrings, write_readme
 
     write_readme(__doc__)
     test_docstrings(__file__)
