@@ -4,8 +4,8 @@ Created on 6. des. 2018
 @author: pab
 """
 
-from __future__ import absolute_import
 from nvector.plot import plot_mean_position
+
 # import sys
 # import pytest
 #
